Refactor crypto utilities and migrate curve functions
- Removed the old crypto utility functions and replaced them with a new centralized crypto module.
- Moved curve-related functions to a dedicated curve module, including key generation, agreement, and signature calculations.
- Updated all references to the new crypto module across the codebase, ensuring consistent usage of the new structure.
- Enhanced error handling and validation in the curve functions.
- Implemented AES encryption and decryption utilities in a separate module for better organization.
- Updated session management and group cipher functionalities to utilize the new crypto methods.
