+++ src/Signal/Group/sender-chain-key.ts
@@ -1,4 +1,4 @@
-import { calculateMAC } from '../Core/crypto'
+import { calculateMAC } from '../../crypto'
 import { SenderMessageKey } from './sender-message-key'
 
 export class SenderChainKey {
@@ -33,6 +33,6 @@ export class SenderChainKey {
 	}
 
 	private getDerivative(seed: Uint8Array, key: Buffer): Uint8Array {
-		return calculateMAC(key, seed)
+		return calculateMAC(key, Buffer.from(seed))
 	}
 }
