+++ src/Utils/index.ts
@@ -3,7 +3,6 @@ export * from './decode-wa-message'
 export * from './messages'
 export * from './messages-media'
 export * from './validate-connection'
-export * from './crypto'
 export * from './signal'
 export * from './noise-handler'
 export * from './history'
