+++ src/Utils/noise-handler.ts
@@ -1,9 +1,9 @@
 import { Boom } from '@hapi/boom'
 import { proto } from '../../WAProto'
+import { aesDecryptGCM, aesEncryptGCM, Curve, hkdf, sha256 } from '../crypto'
 import { NOISE_MODE, WA_CERT_DETAILS } from '../Defaults'
 import { KeyPair } from '../Types'
 import { BinaryNode, decodeBinaryNode } from '../WABinary'
-import { aesDecryptGCM, aesEncryptGCM, Curve, hkdf, sha256 } from './crypto'
 import { ILogger } from './logger'
 
 const generateIV = (counter: number) => {
