+++ src/crypto/aes.ts
@@ -1,0 +1,71 @@
+// AES encryption/decryption utilities
+
+import { createCipheriv, createDecipheriv, randomBytes } from 'crypto'
+
+const GCM_TAG_LENGTH = 128 >> 3
+
+export function aesEncryptGCM(plaintext: Uint8Array, key: Uint8Array, iv: Uint8Array, additionalData: Uint8Array) {
+	const cipher = createCipheriv('aes-256-gcm', key, iv)
+	cipher.setAAD(additionalData)
+	return Buffer.concat([cipher.update(plaintext), cipher.final(), cipher.getAuthTag()])
+}
+
+export function aesDecryptGCM(ciphertext: Uint8Array, key: Uint8Array, iv: Uint8Array, additionalData: Uint8Array) {
+	const decipher = createDecipheriv('aes-256-gcm', key, iv)
+	const enc = ciphertext.slice(0, ciphertext.length - GCM_TAG_LENGTH)
+	const tag = ciphertext.slice(ciphertext.length - GCM_TAG_LENGTH)
+	decipher.setAAD(additionalData)
+	decipher.setAuthTag(tag)
+	return Buffer.concat([decipher.update(enc), decipher.final()])
+}
+
+export function aesEncryptCTR(plaintext: Uint8Array, key: Uint8Array, iv: Uint8Array) {
+	const cipher = createCipheriv('aes-256-ctr', key, iv)
+	return Buffer.concat([cipher.update(plaintext), cipher.final()])
+}
+
+export function aesDecryptCTR(ciphertext: Uint8Array, key: Uint8Array, iv: Uint8Array) {
+	const decipher = createDecipheriv('aes-256-ctr', key, iv)
+	return Buffer.concat([decipher.update(ciphertext), decipher.final()])
+}
+
+// AES CBC with IV prefixed
+export function aesDecrypt(buffer: Buffer, key: Buffer) {
+	return aesDecryptWithIV(buffer.slice(16), key, buffer.slice(0, 16))
+}
+
+function aesDecryptWithIV(buffer: Buffer, key: Buffer, IV: Buffer) {
+	const aes = createDecipheriv('aes-256-cbc', key, IV)
+	return Buffer.concat([aes.update(buffer), aes.final()])
+}
+
+export function aesEncrypt(buffer: Buffer | Uint8Array, key: Buffer) {
+	const IV = randomBytes(16)
+	const aes = createCipheriv('aes-256-cbc', key, IV)
+	return Buffer.concat([IV, aes.update(buffer), aes.final()])
+}
+
+// CBC legacy for Signal protocol
+function assertBuffer(value: any) {
+	if (!(value instanceof Buffer)) {
+		throw TypeError(`Expected Buffer instead of: ${value.constructor.name}`)
+	}
+
+	return value
+}
+
+export function signalEncrypt(key: Buffer, data: Buffer, iv: Buffer) {
+	assertBuffer(key)
+	assertBuffer(data)
+	assertBuffer(iv)
+	const cipher = createCipheriv('aes-256-cbc', key, iv)
+	return Buffer.concat([cipher.update(data), cipher.final()])
+}
+
+export function signalDecrypt(key: Buffer, data: Buffer, iv: Buffer) {
+	assertBuffer(key)
+	assertBuffer(data)
+	assertBuffer(iv)
+	const decipher = createDecipheriv('aes-256-cbc', key, iv)
+	return Buffer.concat([decipher.update(data), decipher.final()])
+}
