+++ src/crypto/signal-protocol.ts
@@ -1,9 +1,8 @@
-// @ts-nocheck
+// Signal protocol-specific cryptographic primitives
 
-import assert from 'node:assert/strict'
 import nodeCrypto from 'node:crypto'
 
-function assertBuffer(value) {
+function assertBuffer(value: any) {
 	if (!(value instanceof Buffer)) {
 		throw TypeError(`Expected Buffer instead of: ${value.constructor.name}`)
 	}
@@ -11,23 +10,7 @@ function assertBuffer(value) {
 	return value
 }
 
-function encrypt(key, data, iv) {
-	assertBuffer(key)
-	assertBuffer(data)
-	assertBuffer(iv)
-	const cipher = nodeCrypto.createCipheriv('aes-256-cbc', key, iv)
-	return Buffer.concat([cipher.update(data), cipher.final()])
-}
-
-function decrypt(key, data, iv) {
-	assertBuffer(key)
-	assertBuffer(data)
-	assertBuffer(iv)
-	const decipher = nodeCrypto.createDecipheriv('aes-256-cbc', key, iv)
-	return Buffer.concat([decipher.update(data), decipher.final()])
-}
-
-function calculateMAC(key, data) {
+export function calculateMAC(key: Buffer, data: Buffer) {
 	assertBuffer(key)
 	assertBuffer(data)
 	const hmac = nodeCrypto.createHmac('sha256', key)
@@ -35,9 +18,7 @@ function calculateMAC(key, data) {
 	return Buffer.from(hmac.digest())
 }
 
-// Salts always end up being 32 bytes
-function deriveSecrets(input, salt, info, chunks = 3) {
-	// Specific implementation of RFC 5869 that only returns the first 3 32-byte chunks
+export function deriveSecrets(input: Buffer, salt: Buffer, info: Buffer, chunks = 3) {
 	assertBuffer(input)
 	assertBuffer(salt)
 	assertBuffer(info)
@@ -45,7 +26,10 @@ function deriveSecrets(input, salt, info, chunks = 3) {
 		throw new Error('Got salt of incorrect length')
 	}
 
-	assert(chunks >= 1 && chunks <= 3)
+	if (chunks < 1 || chunks > 3) {
+		throw new Error('Chunks must be between 1 and 3')
+	}
+
 	const PRK = calculateMAC(salt, input)
 	const infoArray = new Uint8Array(info.byteLength + 1 + 32)
 	infoArray.set(info, 32)
@@ -66,7 +50,7 @@ function deriveSecrets(input, salt, info, chunks = 3) {
 	return signed
 }
 
-function verifyMAC(data, key, mac, length) {
+export function verifyMAC(data: Buffer, key: Buffer, mac: Buffer, length: number) {
 	const calculatedMac = calculateMAC(key, data).slice(0, length)
 	if (mac.length !== length || calculatedMac.length !== length) {
 		throw new Error('Bad MAC length')
@@ -77,4 +61,17 @@ function verifyMAC(data, key, mac, length) {
 	}
 }
 
-export { deriveSecrets, decrypt, encrypt, calculateMAC, verifyMAC }
+// HMAC signing utility (migrated from old Utils/crypto.ts)
+
+export function hmacSign(
+	buffer: Buffer | Uint8Array,
+	key: Buffer | Uint8Array,
+	variant: 'sha256' | 'sha512' = 'sha256'
+) {
+	return nodeCrypto.createHmac(variant, key).update(buffer).digest()
+}
+
+// SHA-256 hash utility (migrated from old Utils/crypto.ts)
+export function sha256(buffer: Buffer) {
+	return require('crypto').createHash('sha256').update(buffer).digest()
+}
